Activated all sensible compiler warnings.
Made sure that all sensible compiler warnings are activated and are
treated as errors. This should lead to higher quality code. The only
warning that has been explicitely disabled is unused-parameter, which
warns about function arguments that are not referenced in the function.
Since unused function arguments are completely valid, for example in
callbacks or the main function, I don't think this type of warning
makes sense.