Prevent SIGPIPE.
If a send call fails due to a broken pipe, the process will be sent a
SIGPIPE, which results in process termination by default. Since I'm
properly checking error codes this is worse than useless.
Fortunately, the SIGPIPE signal can be completely prevented by
providing send with the proper flag, which is what this commit is
doing.