Updated .gitignore.
Configured git to ignore the local VM image I created, as well as the
file that Nomad saves the PID of the VM process in.
Ideally, I would add the VM image to version control (or some other
kind of repository), so it can easily be used by other developers, but
the VM image is almost 2GB big. I have a solution in mind that doesn't
require adding the image to version control, but that would be too
involved for now.