Separated compile and run.
Renamed scripts/helper/run to scripts/helper/compile and moved the
binary invocation out of there. The server and client scripts still
call the helper script to compile, but run the binary themselves.
This makes scripts/helper/compile much more flexible and is a
pre-requisite for running the server inside a local VM.