Implemented receiveMessages.
This is only an initial version that satisfies the current test. It is
not complete yet, as indicated by the several documented edge cases
that have yet to be implemented.
I also think the design is not very good. Maybe just adding all
messages to a single buffer and separate them with scanf would be
simpler.