Encoded message length in first byte.
The length of a string sent by the server has been added as the first
byte of the message. This makes the correct handling of such a message
much easier than it has been previously.
The client code has been amended to ignore the first byte, not making
use of it yet.