Refactoring.
I'm starting to move parts of the array that is used in the client to
hold position into a separate header in a reusable form.
The goal is create a generic data structure that can also be used on
the server side.

In this first step I moved the definition of the entry type.