Refactoring.
Changed posMap into a struct that contains both an array and that
array's length. The array is now allocated on the heap instead of the
stack.

This is another step towards turning idmap into a fully generic data
structure. Allocating the array on the heap makes things much simpler
in that regard.