Remove disconnected clients.
Clients that are no longer connected to the server are removed and
their ids are put back into the pool.
Because the fact that the client has been removed server-side is not
yet sent to the clients in any way, the matching triangle is still
displayed, but it's no longer moving.