Set compiler mode to c1x.
The default standard mode id gnu99, which doesn't support a feature I'd
like to use. This feature is being able use a variable as basis for the
length of a static array instead of having to use a #define.
POSIX functionality doesn't seem to be available in c1x or c99 mode, so
the network code failed to compile. This was fixed by adding the
-D_POSIX_SOURCE bit.