Fixed: Ids were added to pool  multiple times.
It's possible that the clients_remove function is called multiple times
for the same client. If this happened, the client's id was added back
to the pool once for each time the function was called. This could lead
to all kinds of odd behavior and a crash, if the id pool was filled up.