Client fails after receiving unknown message.
Until now, the client would ignore unknown messages. This was done for
forward compatibility, but what it actually does is corrupting the
client state in case of an error and possibly even masking that error.
As of this commit, the client will crash if it receives and unknown
type of message.