Fixed: Client reading partial message.
The condition the client checked for when deciding whether it has a
complete message in its buffer seems wrong. Unfortunately I can't
verify this without better automated testing infrastructure in place.
An issue towards that end has been filed.

So what I'm saying is, I suspect this is a bug, I suspect this commit
is fixing it and I hope I'm not actually creating a new bug with this.