Increased client-side positions limit.
Increased the limit from 2 to 4 to match the number of clients the
server does allow. As long as both values are in sync, the client can't
crash because it received ids that are too large.

Hint: This should probably be defined in a common header file, but the
limit is so low right now that I'm confident I'll run into it again
before this sees any serious use.