Updated GOALS.
Now that the current goal is reached and the networking code functions
somewhat stable, I think it's time to concentrate on graphics. The next
goal is to add a camera that can be rotated around the origin. This can
start be a start to the transition to 3D graphics and gameplay.