Updated GOALS.
The current goal has been reached. Since with the current camera
position, the triangle can't always be seen that well and implementing
a full-blown camera system would be too much of a detour right now
(with little benefit), I think it's time to move the triangle onto an
orbit around the origin.