Refactoring.
Started replacing the fix data type with long. This is the first step
in a series of refactorings that will completely remove the fix type,
first replacing it with long, then replacing long with double.

I still believe that fixed-point arithmetic is the best solution for
the problem at hand (huge solar-system wide coordinate system), but
implementing it is proving to be a challenge. In addition, the lack of
operator overloading in C makes it verbose to use.

If it was possible to implement this without too much trouble, it would
be worth it. As it stands, it's too difficult, doesn't provide any
short-term benefit and thus should be deferred until a later point.