Fixed: scripts/server quits unexpectedly.
scripts/server used to quit under certain circumstances, most commonly
when doing bulk updates using git (like a git stash or git stash pop).
This reason for this was that the loop was set up in such a way that it
would only continue when inotifywait returned 0. inotifywait can return
1 under some completely legitimate circumstances, however, which was
the cause of this bug.

Fixes #20.