Fixed Valgrind error.
When copying a buffer with memcpy, the source and destination
overlapped. This is valid when using memmove, but not with memcpy.

I think this may have been responsible for issues #26, #27 and #28. I
haven't been able to reproduce those since my upgrade to 13.04 and this
Valgrind error seems like a good candidate. It's from the same part of
the code that crashed the client and seems like something that can be
easily influenced by an upgrade.