Refactor
This commit renames Vec3 to Vec2. Of course, this is a really dumb name
for a 3-dimensional vector structure, but the following commits will
rectify that.

I've decided to make this game completely 2-dimensional for the time
being. It will development so much easier and I believe it will be
possible to re-evaluate this decision later.