Remove client-ui's dependence on client-core
Inlined all the functionality which was previously handled by
client-core. As it turns out, that whole thing wasn't that good of an
idea. Client-core is so thin a layer, that the acceptance tests could
just as well connect to the server directly. For client-ui, the whole
thing only meant more work and complexity.

New plan:
- Delete client-core
- Implement alternate (CLI-based) control mode for client-ui
- Hook acceptance tests with client-ui using that control mode

Plan B (in case it doesn't work as I imagine it):
- Extract as much client code as possible into a new libclient
- Use that in both client-ui and the test suite

I hope it won't come to that though. Testing the real client just with
alternate input/output is much better.