Replace network EventBuffer with channel
I'm going to do this with the other EventBufferS, too. Channels allow
much more flexibility (like splitting the components into tasks), are
safer (can't accidentally receive something you're not supposed to) and
if there are ever performance problems that require us to implement a
disruptor, then the interface will be similar to channels anyway.