Use shaders for UI overlay
This replaces the fixed-function pipeline for the rendering of the UI
overlay with a shader-based approach. The way this is done here works
but is extremly slow, as the shader is loaded and compiled every frame.
I've done it this way to contain the changes to a single function. Next:
Cleaning it all up!