Changed the way headless input handler works
Previously, it would send default input data, if no new input was
received. Now, it will repeat the last input until new data is received.

I'm about to change the way input sending works: The client will send
input to the server in regular intervals instead of waiting for the
command. The previous behavior of the headless input handler wouldn't
have worked well with this.