Delete server-side ecs module
The struct it contained has been inlined into module game.
Generic ECS infrastructure belongs clearly in libcommon. That leaves
only components for the game service's ecs module at the moment, and I
don't think there's anything ECS-specific about them. A component is
just a generic data structure and the only thing that ties it to the ECS
is an entity.