Create trait for entity templates
This trait will help to consolidate some code in a central place that
would otherwise be duplicated. It will also validate entities with the
help of the type system as much as possible.

Unfortunately, the current trait only works for entities with two
components. For each number of components, a separate trait will have to
be added. There's nothing to be done about this, since Rust doesn't
support variadic type arguments at the moment.