Upgrade Rust
This includes only the upgrade to the latest Rust version, not any of
the modifications that upgrade requires. Therefore, this commit breaks
the build.

This is not ideal, but I don't have time to finish the upgrade tonight.
I think it's better to commit whatever I have finished so far, even
though that breaks the build for now.