Pass previous snapshot into Perception::new
Perception is about to grow new capabilities, namely it will be possbile
to distinguish between created, destroyed and updated entities. The
current behavior (all entities are sent as updated) can be preserved by
passing identical entities for both arguments.

This commit changes a call to Perception::new in such a way that it
won't be affected by the new features.