Change the way ECS import functions work
For every entity there's now an import function in addition to a create
function. The only difference between them is that the import function
accepts an id.

The generation code for the import function is mostly the same as the
code for the create function. The ideal solution would be, if the
create function would just generate the id and then call the import
function. However, there's a problem with that relating to the
arguments. It seems quite difficult to get the arugment names out of the
Vec<ast::Arg> that we have available.