Improve craft display
This replaces the billboarding hack with something more robust, which
makes the craft icon and text look much better.

Please note that zooming doesn't work correctly, which means the grid
and crafts seem to zoom differently. I confirmed that this was the case
with the old hack, too, so this isn't a regression in this commit.