Move args module to server lib
I'm beginning the process of moving most client/server code into the
libs. There are a few reasons for this:
a) The separation seems artificial. It's not clear where to look for a
   given piece of functionality.
b) When I need to access something from the tests, I need to move it
   first. This friction can be removed by having everything in the libs
   in the first place.

I don't believe this will make a difference to build times, since for
the common cases, all the same stuff still needs to be built.