Move client binary into dedicated package
I believe moving the binaries into dedicated packages will remove the
cause of the build failure I keep seeing from time to time. I think that
failure is caused by the client and server libraries being built twice
each, once directly and once as a dependency of the tests.

I also hoped it would improve build times, but it actually seems to have
an adverse effect. I'll keep a look on this. Maybe it'll be worth going
the exactly opposite route of moving everything into a single package.