Count number of chars to center text instead
String::len returns the number of bytes in the string, not the number of
chars. Both are the same for the subset of characters used in English
and many other languages, but this turns into a bug as soon as you have
a multi-byte characters in your string, for example anything Chinese.

@viperscape:
As a general rule, String::len is almost always not what
you actually want to use :)
I found out the hard way (not in Rust, but the issue exists in any
language that uses UTF-8 string).