Fix spinning connection thread in server
When a remote socket had been closed, the server would not always
recognize this, and the connection thread would spin in an endless loop.

This happened under the following circumstances:
1. No message had been received for a while from the client, causing it
   to be removed after the timeout period.
2. However, the socket was still open, so the client was still present
   in the network layer.
3. The client then closed the socket.
4. Since the client only lived on in the network layer, no messages were
   sent to it. Consequently, its absence would not be detected by
   failing to send a message.

This commit fixes the error detection when receiving messages, making
this problem disappear. I'm not sure why, but I was originally under the
impression that getting an empty line from the socket from time to time
was normal and didn't indicate an unrecoverable error.