Improve legibility by adding some whitespace
@viperscape: I've noticed you often write these very compact if
expressions. While I think that having the whole expression in a single
line can be fitting for very short if expressions, cramming whole
if-else expressions into a few lines makes them quite hard to read. I
much prefer the more spacious style, as condition, if-branch and
else-branch are clearly demarcated.