Add Ship component to entity system
I believe the currently existing attributes component is not really in
the spirit of an entity system. What `Attributes` basically does is say
"this entity is of type ship/planet", but entities don't really have a
"type", in the traditional object-oriented sense. The type of an entity
is defined by the components it has.

I believe it's more fitting to mark the type of a "ship" component by
adding the component `Ship`, instead of using `Attributes` for this.
Right now, `Ship` is empty and acts purely as a marker, but I expect us
to add data there, as the game grows.

Our entity system implementation currently lacks support for this, but
eventually it will be possible to iterate over entities by their
components (e.g. "iterate over all entities with `Body`, `Ship`). This
makes the approach with marker components more natural then iterating
over all bodies and checking attributes for the type.

cc @viperscape