Add support for planet component to entity system
This is supposed to replace Attributes in its entirety. It is a marker
component, similar to `Ship`, but also contains the planet data, just as
`Ship` will likely contain data in the future.

Currently the fields of `Planet` duplicate `PlanetAttr`, but I'm going
to merge those two together in the subsequent commits.

cc @viperscape