Add set of all entities to Entities
So far, it was only possible to iterate over components. This is good
most of the time, but it's not good enough for exporting entities in the
general case. This has worked up till now, since all exported entities
had a `Body`.

I'm about to export maneuver entities, too. Since those don't have a
body, iterating over all bodies for exporting won't cut it any more.