Make communication of entity removal more specific
As of now, the only cause of entity removal is a player leaving the
game. This has been handled in the main loop. However, in the future, we
want to communicate other removal events (destroyed ships, expired
maneuvers) to the client. This needs a generic way for the game state to
communicate which entities have been destroyed, which this commit
implements.