Add intermediate representation for initial state
Instead of randomly generating the entities directly, an immediate
representation of the initial state is generated. The entities are then
created from that immediate representation.

This has the advantage that the intermediate representation can come
from different sources. This means, it can be randomly generated in
production use, or it can be tightly controlled by test code in a test
scenario.