Remove unused code
All that was needed for ship-planet collision was a simple point-sphere
intersection check. The collider component wasn't needed there, and I
don't think we'll need sphere-sphere collision detection for the
forseeable future.

This commit also removes the collision event that the server sent to the
client. It wasn't used to do anything there, and the client already
knows the ship was removed via the RemoveEntity event.