Set default log level for test runs to debug
As it turns out, the Rust compiler has its own logging crate that seems
to duplicate the functionality of env_logger, but doesn't support the
trace level. Since it reads the same environment variable, it prints out
warnings on each run, which amounts to a lot of noise during a typical
cargo run.