Fix crash on mouse button release without press
If the mouse button was released without having been pressed first, the
game would panic on an unwrap. Releasing without pressing can happen, if
the mouse button was pressed without an event being sent to the window.
This was something I could reliably reproduce by maximising the window
via a double-click on the title bar.